Complete the last argument if not matching a hint
Instead of stopping the completion when there is a value for an
argument, try to verify if the value is matching one of the valid
option. If it's not matching, try to suggest all options starting by the
entered value.

(cherry picked from commit 42d15048d91418a68cb7c837a77c99fc8d56b1ac)