Add possibility to hide arguments from usage/help
This is for example useful when a command accepts an argument for
backwards compatability purposes only. As in this case the argument
is no longer the recommended way to do things, it should not be
included in the usage output.

(cherry picked from commit 0a108b7f5563f8b34535f13cbb0b811dce75b87c)