Fix autocompletion of args that consume remainder
The Go runtime would panic at line 35 because of an out-of-bounds
slice access if autocompletion was requested for a clause that
consumed the remainder of a command. By specifically checking for
these kinds of clauses, the problem should be mitigated.

(cherry picked from commit cf44c60659e38b5559aa885c13ad76c9bd6e8d50)