Honour -- delimiter in autocompletion
If a -- argument is passed, from then on only arguments should be
parsed. This is already handled correctly in most places, but
autocompleting did not honour this yet.

Setting `p.argsOnly = true` was moved to the next call of Next()
because otherwise a completion for a flag was incorrectly treated
as a an `--` argument, causing the completion code not to return
any suggestions for flags.

(cherry picked from commit 5d6b81cf32640d71099a7a731406bcf230c05348)